+++ swift/Sources/Cindermark/CindermarkFFI.swift
@@ -1947,6 +1947,7 @@ public enum FfiBlockType {
     )
     case math(syntax: UInt8, quoteDepth: UInt32, listKind: UInt8, listContentIndent: UInt32, contentUtf16Start: UInt32, contentUtf16End: UInt32
     )
+    case rawHtml
 }
 
 
@@ -1996,6 +1997,8 @@ public struct FfiConverterTypeFfiBlockType: FfiConverterRustBuffer {
         
         case 17: return .math(syntax: try FfiConverterUInt8.read(from: &buf), quoteDepth: try FfiConverterUInt32.read(from: &buf), listKind: try FfiConverterUInt8.read(from: &buf), listContentIndent: try FfiConverterUInt32.read(from: &buf), contentUtf16Start: try FfiConverterUInt32.read(from: &buf), contentUtf16End: try FfiConverterUInt32.read(from: &buf)
         )
+        
+        case 18: return .rawHtml
         
         default: throw UniffiInternalError.unexpectedEnumCase
         }
@@ -2080,6 +2083,10 @@ public struct FfiConverterTypeFfiBlockType: FfiConverterRustBuffer {
             FfiConverterUInt32.write(contentUtf16Start, into: &buf)
             FfiConverterUInt32.write(contentUtf16End, into: &buf)
             
+        
+        case .rawHtml:
+            writeInt(&buf, Int32(18))
+        
         }
     }
 }